Refactor item selection and update file paths in ItemCreationPage.vue
The code for selecting month in `ItemCreationPage.vue` was moved to `App.vue`, and prop 'passed_month' is now required. The paths for items.json and item icon files were updated, and the 'watch' function was added to monitor 'selected_month' changes. Some adjustments were made in textBoxes formatting and a selected_large_number field is now included in the item data.